+++ cli/src/pubsub.rs
@@ -10,6 +10,7 @@ use {
     solana_account_decoder::{UiAccountEncoding, UiDataSliceConfig},
     solana_client::nonblocking::pubsub_client::PubsubClient,
     solana_commitment_config::CommitmentConfig,
+    solana_pubkey::Pubkey,
     solana_rpc_client_api::{
         config::{
             RpcAccountInfoConfig, RpcBlockSubscribeConfig, RpcBlockSubscribeFilter,
@@ -19,7 +20,7 @@ use {
         filter::{Memcmp, MemcmpEncodedBytes, RpcFilterType},
         response::RpcVersionInfo,
     },
-    solana_sdk::{pubkey::Pubkey, signature::Signature},
+    solana_signature::Signature,
     solana_transaction_status::{TransactionDetails, UiTransactionEncoding},
     std::{collections::HashMap, fmt, str::FromStr},
     tokio::{